slowly moving away from step
